+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add support for `cognitor` and `cognitor-worker` log level selection from `cognitor`'s `.env` file.
 - Added support for `.log` files ingestion.
+- Added support for ingestion status tracking
 
 ## Release v0.1.0 - June 10, 2026